replace 'org' var for 'git' replace 'hub' command for 'git' play the reset/pull only if target is a git repository

+++ scripts/git-get.js
@@ -1,11 +1,11 @@
 var shell = require('shelljs');
 var fs = require('fs-extra');
-var type = process.argv[2] || false;
-var name = process.argv[3] || false;
-var org  = process.argv[4] || false;
+var type 	= process.argv[2] || false;
+var name 	= process.argv[3] || false;
+var git  	= process.argv[4] || 'https://github.com/SeptimusVII/';
 
 if(!type || !name){
-  console.log('\n Missing argument(s). Expected syntax: git-get [type] [name] \n');
+  console.log('\nMissing argument(s). Expected syntax: git-get [type] [name] \n');
 } else {
 	if (!fs.existsSync('./src/'+type+'s/'+name+'/')){
 	    fs.mkdir('./src/'+type+'s/'+name+'/',function(err){
@@ -13,15 +13,16 @@ if(!type || !name){
 	        console.log('\n'+err.message+'\n');
 	    });
 		shell.cd('./src/'+type+'s/');
-		if(!org)
-			shell.exec('hub clone framway-'+type+'-'+name+' '+name);
-		else
-			shell.exec('hub clone '+org+'/framway-'+type+'-'+name+' '+name);
+		shell.exec('git clone '+git+'framway-'+type+'-'+name+' '+name);
 		shell.cd();
 	} else {
-		shell.cd('./src/'+type+'s/'+name);
-		shell.exec('git reset --hard');
-		shell.exec('git pull --rebase');
-		shell.cd();
+		if (fs.existsSync('./src/'+type+'s/'+name+'/.git')){
+			shell.cd('./src/'+type+'s/'+name);
+			shell.exec('git reset --hard');
+			shell.exec('git pull --rebase');
+			shell.cd();
+		} else {
+			console.log('Error: '+name+' has no git repository \n');
+		}
 	}
 }
